+++ Hotel1/user-interface/src/main/java/com/senlainc/git_courses/java_training/petushok_valiantsin/user_interface/ui/order/AddOrder.java
@@ -23,10 +23,10 @@ public class AddOrder implements IAction {
             final Scanner scanner = new Scanner(System.in);
             hotel.showGuest().forEach(System.out::println);
             System.out.print("Enter guest index: ");
-            final int guestIndex = Integer.parseInt(scanner.nextLine());
+            final long guestIndex = Long.parseLong(scanner.nextLine());
             hotel.showRoom("all").forEach(System.out::println);
             System.out.print("Enter room index: ");
-            final int roomIndex = Integer.parseInt(scanner.nextLine());
+            final long roomIndex = Long.parseLong(scanner.nextLine());
             System.out.print("Enter end date(Format: YYYY-MM-DD): ");
             final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
             final LocalDate endDate = LocalDate.parse(scanner.nextLine(), formatter);
